Binary masking and random decision
Included a binary mask for the string mask choice. The binary mask selects only non-white (string) pixels and includes them in the string image. This avoids blending the white background when building the final image. The random decision is based on a random variable and gives a certain probability to pick a random nail as the next nail; may help the algorithm not get stuck.